+++ include/common/complex16x2simd.hpp
@@ -1,6 +1,6 @@
 //////////////////////////////////////////////////////////////////////////////////////
 //
-// (C) Daniel Strano and the Qrack contributors 2017, 2018. All rights reserved.
+// (C) Daniel Strano and the Qrack contributors 2017-2019. All rights reserved.
 //
 // This is a SIMD implementation of the double precision complex type.
 // The API is designed to (almost entirely) mirror that of the C++ standard library
