+++ src/qinterface/gates.cpp
@@ -1,6 +1,6 @@
 //////////////////////////////////////////////////////////////////////////////////////
 //
-// (C) Daniel Strano and the Qrack contributors 2017, 2018. All rights reserved.
+// (C) Daniel Strano and the Qrack contributors 2017-2019. All rights reserved.
 //
 // This is a multithreaded, universal quantum register simulation, allowing
 // (nonphysical) register cloning and direct measurement of probability and
@@ -190,6 +190,57 @@ void QInterface::UniformlyControlledSingleBit(
                 X(controls[bit_pos]);
             }
         }
+    }
+}
+
+void QInterface::TimeEvolve(Hamiltonian h, real1 timeDiff)
+{
+    // Exponentiation of an arbitrary serial string of gates, each HamiltonianOp component times timeDiff, e^(-i * H *
+    // t) as e^(-i * H_(N - 1) * t) * e^(-i * H_(N - 2) * t) * ... e^(-i * H_0 * t)
+
+    for (bitLenInt i = 0; i < h.size(); i++) {
+        HamiltonianOpPtr op = h[i];
+        complex* opMtrx = op->matrix.get();
+        complex* mtrx;
+
+        bitCapInt maxJ = 4;
+        if (op->uniform) {
+            maxJ *= 1U << op->controlLen;
+        }
+        mtrx = new complex[maxJ];
+
+        for (bitCapInt j = 0; j < maxJ; j++) {
+            mtrx[j] = opMtrx[j] * (-timeDiff);
+        }
+
+        if (op->toggles) {
+            for (bitLenInt j = 0; j < op->controlLen; j++) {
+                if (op->toggles[j]) {
+                    X(op->controls[j]);
+                }
+            }
+        }
+
+        if (op->uniform) {
+            complex* expMtrx = new complex[maxJ];
+            for (bitCapInt j = 0; j < (1U << op->controlLen); j++) {
+                exp2x2(mtrx + (j * 4), expMtrx + (j * 4));
+            }
+            UniformlyControlledSingleBit(op->controls, op->controlLen, op->targetBit, expMtrx);
+            delete[] expMtrx;
+        } else {
+            Exp(op->controls, op->controlLen, op->targetBit, mtrx, op->anti);
+        }
+
+        if (op->toggles) {
+            for (bitLenInt j = 0; j < op->controlLen; j++) {
+                if (op->toggles[j]) {
+                    X(op->controls[j]);
+                }
+            }
+        }
+
+        delete[] mtrx;
     }
 }
