+++ src/qinterface/logic.cpp
@@ -14,40 +14,6 @@
 
 namespace Qrack {
 
-// Arithmetic:
-/// Subtract integer (without sign)
-void QInterface::DEC(bitCapInt toSub, bitLenInt inOutStart, bitLenInt length)
-{
-    bitCapInt invToSub = (1U << length) - toSub;
-    INC(invToSub, inOutStart, length);
-}
-
-/**
- * Subtract an integer from the register, with sign and without carry. Because the register length is an arbitrary
- * number of bits, the sign bit position on the integer to add is variable. Hence, the integer to add is specified as
- * cast to an unsigned format, with the sign bit assumed to be set at the appropriate position before the cast.
- */
-void QInterface::DECS(bitCapInt toSub, bitLenInt inOutStart, bitLenInt length, bitLenInt overflowIndex)
-{
-    bitCapInt invToSub = (1U << length) - toSub;
-    INCS(invToSub, inOutStart, length, overflowIndex);
-}
-
-/// Subtract integer (without sign, with controls)
-void QInterface::CDEC(
-    bitCapInt toSub, bitLenInt inOutStart, bitLenInt length, bitLenInt* controls, bitLenInt controlLen)
-{
-    bitCapInt invToSub = (1U << length) - toSub;
-    CINC(invToSub, inOutStart, length, controls, controlLen);
-}
-
-/// Subtract BCD integer (without sign)
-void QInterface::DECBCD(bitCapInt toSub, bitLenInt inOutStart, bitLenInt length)
-{
-    bitCapInt invToSub = intPow(10U, length / 4U) - toSub;
-    INCBCD(invToSub, inOutStart, length);
-}
-
 // Logic Gates:
 
 /// "AND" compare two bits in QInterface, and store result in outputBit
@@ -193,54 +159,4 @@ void QInterface::CLXOR(bitLenInt inputQBit, bool inputClassicalBit, bitLenInt ou
     }
 }
 
-void QInterface::TimeEvolve(Hamiltonian h, real1 timeDiff)
-{
-    // Exponentiation of an arbitrary serial string of gates, each HamiltonianOp component times timeDiff, e^(-i * H *
-    // t) as e^(-i * H_(N - 1) * t) * e^(-i * H_(N - 2) * t) * ... e^(-i * H_0 * t)
-
-    for (bitLenInt i = 0; i < h.size(); i++) {
-        HamiltonianOpPtr op = h[i];
-        complex* opMtrx = op->matrix.get();
-        complex* mtrx;
-
-        bitCapInt maxJ = 4;
-        if (op->uniform) {
-            maxJ *= 1U << op->controlLen;
-        }
-        mtrx = new complex[maxJ];
-
-        for (bitCapInt j = 0; j < maxJ; j++) {
-            mtrx[j] = opMtrx[j] * (-timeDiff);
-        }
-
-        if (op->toggles) {
-            for (bitLenInt j = 0; j < op->controlLen; j++) {
-                if (op->toggles[j]) {
-                    X(op->controls[j]);
-                }
-            }
-        }
-
-        if (op->uniform) {
-            complex* expMtrx = new complex[maxJ];
-            for (bitCapInt j = 0; j < (1U << op->controlLen); j++) {
-                exp2x2(mtrx + (j * 4), expMtrx + (j * 4));
-            }
-            UniformlyControlledSingleBit(op->controls, op->controlLen, op->targetBit, expMtrx);
-            delete[] expMtrx;
-        } else {
-            Exp(op->controls, op->controlLen, op->targetBit, mtrx, op->anti);
-        }
-
-        if (op->toggles) {
-            for (bitLenInt j = 0; j < op->controlLen; j++) {
-                if (op->toggles[j]) {
-                    X(op->controls[j]);
-                }
-            }
-        }
-
-        delete[] mtrx;
-    }
-}
 } // namespace Qrack
